test when remove suiviCreatedEvent #5246
