+++ src/Command/Cron/SendSuiviWaitingNotificationCommand.php
@@ -2,10 +2,9 @@
 
 namespace App\Command\Cron;
 
-use App\Event\SuiviCreatedEvent;
+use App\Manager\SuiviManager;
 use App\Repository\SuiviRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Psr\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Style\SymfonyStyle;
@@ -21,7 +20,7 @@ class SendSuiviWaitingNotificationCommand extends AbstractCronCommand
         private readonly SuiviRepository $suiviRepository,
         private readonly EntityManagerInterface $entityManager,
         private readonly ParameterBagInterface $parameterBag,
-        private readonly EventDispatcherInterface $eventDispatcher,
+        private readonly SuiviManager $suiviManager,
     ) {
         parent::__construct($this->parameterBag);
     }
@@ -31,7 +30,7 @@ class SendSuiviWaitingNotificationCommand extends AbstractCronCommand
         $suivis = $this->suiviRepository->findWithWaitingNotificationAndExpiredDelay();
         foreach ($suivis as $suivi) {
             $suivi->setWaitingNotification(false);
-            $this->eventDispatcher->dispatch(new SuiviCreatedEvent($suivi), SuiviCreatedEvent::NAME); // @phpstan-ignore-line
+            $this->suiviManager->onSuiviCreated($suivi);
         }
 
         $this->entityManager->flush();
