+++ src/DataFixtures/Loader/LoadSuiviData.php
@@ -5,14 +5,12 @@ namespace App\DataFixtures\Loader;
 use App\Entity\Enum\SignalementStatus;
 use App\Entity\Enum\SuiviCategory;
 use App\Entity\Suivi;
-use App\Event\SuiviCreatedEvent;
 use App\Manager\SuiviManager;
 use App\Repository\SignalementRepository;
 use App\Repository\UserRepository;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
 use Doctrine\Persistence\ObjectManager;
-use Psr\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use Symfony\Component\Yaml\Yaml;
 
@@ -23,7 +21,6 @@ class LoadSuiviData extends Fixture implements OrderedFixtureInterface
         private readonly UserRepository $userRepository,
         private readonly ParameterBagInterface $parameterBag,
         private readonly SuiviManager $suiviManager,
-        private readonly EventDispatcherInterface $eventDispatcher,
     ) {
     }
 
@@ -99,7 +96,7 @@ class LoadSuiviData extends Fixture implements OrderedFixtureInterface
         );
         if (isset($row['force_notifications']) && $row['force_notifications']) {
             $suivi->setWaitingNotification(false);
-            $this->eventDispatcher->dispatch(new SuiviCreatedEvent($suivi), SuiviCreatedEvent::NAME); // @phpstan-ignore-line
+            $this->suiviManager->onSuiviCreated($suivi);
         }
 
         $manager->persist($suivi);
